Optimize sum-coverage (#159)
We'd like to avoid looping through a report's source files each time we add a new source file.

The reason I needed to change the test is because this implementation only works if only only add source files via the AddSourceFile method... (which is the case in the sum-coverage command code)

For a large data set, this shaves the sum-coverage time from 24m to 4m (on my laptop).

Also, this may be a "drive by", but I added some error-handling code in the call to report.Merge; I was hitting this in my initial benchmarking, and thought it was succeeding quickly in some cases, but it was actually failing quickly. We're planning to remove this command's dependence on git, so maybe I should just remove that commit?