Treat 409 upload status as warning, not error
Depending on CI setup, it is possible that CI can run multiple times for
the same commit. The `v1/test_reports` endpoint returns 409 when
receiving a request to create a test report when one already exists for
the commit. This exception currently is bubbled through the reporter
causing the run to ultimately exit with status 1.

This changes the behavior to treat the 409 status separately as a
"warning" rather than an "error," which better reflects the situation.
The exit status in this case will be 0 and a warning will be logged.

Another alternative considered was changing the endpoint to be
create-or-update, but that has been deferred for now.